【改善】ステータスの値を float に変更

+++ Assets/Project/Scripts/Characters/DamageCalculator.cs
@@ -2,9 +2,9 @@ using System;
 using UnityEngine;
 
 public static class DamageCalculator {
-    public static int Calculate(Character attacker, Character target) {
-        int atk = attacker.GetFinalStatus(StatusType.PhysicalAttack);
-        int def = target.GetFinalStatus(StatusType.Defense);
+    public static float Calculate(Character attacker, Character target) {
+        float atk = attacker.GetFinalStatus(StatusType.PhysicalAttack);
+        float def = target.GetFinalStatus(StatusType.Defense);
 
         return Mathf.Max(1, atk - def);
     }
