+++ Assets/Project/Scripts/Effects/StatusBonusEffect.cs
@@ -7,14 +7,14 @@ public class StatusBonusEffect : Effect {
     [Serializable]
     public class StatusEntry {
         public StatusType statusType;
-        public int value;
+        public float value;
     }
 
     [SerializeField] private List<StatusEntry> _status = new();
 
     // 強化アイテムによるステータス変更を反映
-    public override int GetStatusBonus(StatusType type) {
-        int total = 0;
+    public override float GetStatusBonus(StatusType type) {
+        float total = 0;
 
         foreach (var s in _status) {
             if (s.statusType == type) total += s.value;
